apps/bms-iot-app/pyproject.toml
[project]
name = "bms-iot-app"
version = "0.1.0"
description = "BMS IoT Application with BACnet support"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "BAC0==2025.06.10",
    "bacpypes>=0.19.0",
    "httpx==0.25.2",
    "paho-mqtt>=2.1.0",
    "pydantic==2.5.3",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "trio==0.30.0",
    "typer>=0.9.0",
    "trio-asyncio==0.15.0",
    "pykka==4.2.0",
    "sqlmodel>=0.0.8",
    "loguru==0.7.3"
]
requires-python = ">=3.11"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.1",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-isolate",
    "memory-profiler>=0.61.0",
    "psutil>=5.9.0",
]

[project.scripts]
bms-iot = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --asyncio-mode=auto"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "isolate: mark test to run in isolated subprocess",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
target-version = "py311"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
